remove Newtonsoft related deprecated functions in JSON and replace them with System.Text.Json variants
