+++ src/Serenity.Net.Entity/FieldTypes/JsonField.cs
@@ -108,8 +108,8 @@ public class JsonField<TValue> : GenericClassField<TValue>
                 return 1;
         }
         else
-            return JSON.Serialize(value1, writeNulls: false)
-                .CompareTo(JSON.Serialize(value2, writeNulls: false));
+            return JSON.Stringify(value1, writeNulls: false)
+                .CompareTo(JSON.Stringify(value2, writeNulls: false));
     }
 
     /// <summary>
@@ -142,12 +142,8 @@ public class JsonField<TValue> : GenericClassField<TValue>
                 _setValue(row, null);
                 break;
             case Newtonsoft.Json.JsonToken.String:
-                if (typeof(TValue) == typeof(string))
-                    _setValue(row, serializer.Deserialize<TValue>(reader));
-                else
-#pragma warning disable CS0618 // Type or member is obsolete
-                    _setValue(row, JSON.Parse<TValue>((string)reader.Value, includeNulls: serializer.NullValueHandling == Newtonsoft.Json.NullValueHandling.Include));
-#pragma warning restore CS0618 // Type or member is obsolete
+                _setValue(row, Newtonsoft.Json.JsonConvert.DeserializeObject<TValue>(
+                    (string)reader.Value, JsonSettings.StrictIncludeNulls));
                 break;
             default:
                 _setValue(row, serializer.Deserialize<TValue>(reader));
