+++ src/Serenity.Net.Web/Upload/MultipleFileUploadBehavior.cs
@@ -82,7 +82,7 @@ public class MultipleFileUploadBehavior : BaseSaveDeleteBehavior, IImplicitBehav
             !json.EndsWith("]", StringComparison.Ordinal)))
             throw new ArgumentOutOfRangeException(key);
 
-        var list = JSON.Deserialize<UploadedFile[]>(json ?? "[]");
+        var list = JSON.Parse<UploadedFile[]>(json ?? "[]");
 
         if (list.Any(x => string.IsNullOrEmpty(x.Filename)) ||
             list.GroupBy(x => x.Filename.Trim()).SelectMany(x => x.Skip(1)).Any())
@@ -205,7 +205,7 @@ public class MultipleFileUploadBehavior : BaseSaveDeleteBehavior, IImplicitBehav
             file.Filename = copyResult.Path;
         }
 
-        return JSON.Serialize(newFileList, writeNulls: false);
+        return JSON.Stringify(newFileList, writeNulls: false);
     }
 
     /// <inheritdoc/>
